+++ modules/features/filters/src/main/java/au/com/shiftyjelly/pocketcasts/playlists/manual/AddToPlaylistFragment.kt
@@ -26,8 +26,8 @@ import au.com.shiftyjelly.pocketcasts.compose.CallOnce
 import au.com.shiftyjelly.pocketcasts.compose.components.AnimatedNonNullVisibility
 import au.com.shiftyjelly.pocketcasts.compose.components.ThemedSnackbarHost
 import au.com.shiftyjelly.pocketcasts.models.to.EpisodeUuidPair
-import au.com.shiftyjelly.pocketcasts.models.to.PlaylistPreviewForEpisode
 import au.com.shiftyjelly.pocketcasts.playlists.PlaylistFragment
+import au.com.shiftyjelly.pocketcasts.playlists.manual.AddToPlaylistViewModel.PlaylistChangeFeedback
 import au.com.shiftyjelly.pocketcasts.repositories.playlist.Playlist
 import au.com.shiftyjelly.pocketcasts.ui.helper.FragmentHostListener
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
@@ -160,29 +160,23 @@ internal class AddToPlaylistFragment : BaseDialogFragment() {
     override fun onDismiss(dialog: DialogInterface) {
         super.onDismiss(dialog)
         if (!requireActivity().isChangingConfigurations && isDoneTapped) {
-            showDoneSnackbar(viewModel.getPlaylistsAddedTo())
+            showDoneSnackbar(viewModel.getPlaylistChangeFeedback())
         }
     }
 
-    private fun showDoneSnackbar(playlistsAddedTo: Set<PlaylistPreviewForEpisode>) {
+    private fun showDoneSnackbar(feedback: PlaylistChangeFeedback) {
         val hostListener = requireActivity() as FragmentHostListener
         val snackbarView = hostListener.snackBarView()
-        val snackbar = when (val size = playlistsAddedTo.size) {
-            0 -> return
+        val message = when (feedback) {
+            is PlaylistChangeFeedback.StringResource -> getString(feedback.resourceId)
 
-            1 -> {
-                val playlist = playlistsAddedTo.first()
-                val message = getString(LR.string.added_to_playlist_single, playlist.title)
-                Snackbar.make(snackbarView, message, Snackbar.LENGTH_LONG)
-                    .setAction(LR.string.view) { hostListener.openManualPlaylist(playlist.uuid) }
+            is PlaylistChangeFeedback.PluralResource -> {
+                resources.getQuantityString(feedback.resourceId, feedback.quantity, feedback.quantity)
             }
 
-            else -> {
-                val message = resources.getQuantityString(LR.plurals.added_to_playlist_single_multiple, size, size)
-                Snackbar.make(snackbarView, message, Snackbar.LENGTH_LONG)
-            }
+            PlaylistChangeFeedback.None -> return
         }
-        snackbar.show()
+        Snackbar.make(snackbarView, message, Snackbar.LENGTH_LONG).show()
     }
 
     @Composable