+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 *   New Features
     *   Add option to switch to the HLS video stream when playing a downloaded episode
         ([#5620](https://github.com/Automattic/pocket-casts-android/pull/5620))
+    *   Improve feedback when adding episodes to playlists
+        ([#5632](https://github.com/Automattic/pocket-casts-android/pull/5632))
 *   Bug Fixes
     *   Allow copying description and show notes links with a long press
         ([#5628](https://github.com/Automattic/pocket-casts-android/pull/5628))